Add sessionComm hook and add styling
